Fixed lifetime issues on mqtt payloads, separated out test fixture
