+++ src/test/java/software/amazon/awssdk/crt/test/MqttClientConnectionTest.java
@@ -15,147 +15,10 @@
 
 package software.amazon.awssdk.crt.test;
 
-import org.junit.Assert;
 import org.junit.Assume;
 import org.junit.Test;
-import static org.junit.Assert.*;
+import software.amazon.awssdk.crt.CrtResource;;
 
-import software.amazon.awssdk.crt.*;
-import software.amazon.awssdk.crt.io.ClientBootstrap;
-import software.amazon.awssdk.crt.io.EventLoopGroup;
-import software.amazon.awssdk.crt.io.HostResolver;
-import software.amazon.awssdk.crt.io.TlsContext;
-import software.amazon.awssdk.crt.io.TlsContextOptions;
-import software.amazon.awssdk.crt.mqtt.*;
-
-import java.nio.file.InvalidPathException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Date;
-import java.util.UUID;
-import java.util.concurrent.CompletableFuture;
-
-class MissingCredentialsException extends RuntimeException {
-    MissingCredentialsException(String message) {
-        super(message);
-    }
-}
-
-class MqttClientConnectionFixture {
-
-    MqttClientConnection connection = null;
-    private boolean disconnecting = false;
-
-    static final String TEST_ENDPOINT = System.getProperty("endpoint");
-    static final String TEST_CERTIFICATE = System.getProperty("certificate");
-    static final String TEST_PRIVATEKEY = System.getProperty("privatekey");
-    static final String TEST_ROOTCA = System.getProperty("rootca");
-    static final short TEST_PORT = 8883;
-    static final short TEST_PORT_ALPN = 443;
-    static final String TEST_CLIENTID = "sdk-java-v2-" + UUID.randomUUID();
-
-    Path pathToCert = null;
-    Path pathToKey = null;
-    Path pathToCa = null;
-
-    private boolean findCredentials() {
-        try {
-            pathToCert = Paths.get(TEST_CERTIFICATE);
-            pathToKey = Paths.get(TEST_PRIVATEKEY);
-            pathToCa = Paths.get(TEST_ROOTCA);
-            if (pathToCert == null || pathToCert.toString().equals("")) {
-                throw new MissingCredentialsException("Certificate not provided");
-            }
-            if (!pathToCert.toFile().exists()) {
-                throw new MissingCredentialsException("Certificate could not be found at " + pathToCert);
-            }
-            if (pathToKey == null || pathToKey.toString().equals("")) {
-                throw new MissingCredentialsException("Private key not provided");
-            }
-            if (!pathToKey.toFile().exists()) {
-                throw new MissingCredentialsException("Private key could not be found at " + pathToKey);
-            }
-            if (pathToCa != null && !pathToCa.toFile().exists()) {
-                throw new MissingCredentialsException("Root CA could not be found at " + pathToCa);
-            }
-            return true;
-        } catch (InvalidPathException ex) {
-            return false;
-        } catch (MissingCredentialsException ex) {
-            return false;
-        }
-    }
-
-    MqttClientConnectionFixture() {
-    }
-
-    boolean connect() {
-        return connect(false, 0);
-    }
-
-    boolean connect(boolean cleanSession, int keepAliveMs) {
-        Assume.assumeTrue(findCredentials());
-
-        MqttClientConnectionEvents events = new MqttClientConnectionEvents(){
-            @Override
-            public void onConnectionResumed(boolean sessionPresent) {
-                System.out.println("Connection resumed");
-            }
-
-            @Override
-            public void onConnectionInterrupted(int errorCode) {
-                if (!disconnecting) {
-                    System.out.println("Connection interrupted: error: " + errorCode + " " + CRT.awsErrorString(errorCode));
-                }
-            }
-        };
-
-        try(EventLoopGroup elg = new EventLoopGroup(1);
-            HostResolver hr = new HostResolver(elg);
-            ClientBootstrap bootstrap = new ClientBootstrap(elg, hr);
-            TlsContextOptions tlsOptions = TlsContextOptions.createWithMTLSFromPath(pathToCert.toString(), pathToKey.toString())) {
-
-            int port = TEST_PORT;
-            if (!pathToCa.toString().equals("")) {
-                tlsOptions.overrideDefaultTrustStoreFromPath(null, pathToCa.toString());
-            }
-            if (TlsContextOptions.isAlpnSupported()) {
-                tlsOptions.setAlpnList("x-amzn-mqtt-ca");
-                port = TEST_PORT_ALPN;
-            }
-
-            cleanSession = true; // only true is supported right now
-            String clientId = TEST_CLIENTID + (new Date()).toString();
-            try (TlsContext tls = new TlsContext(tlsOptions);
-                MqttClient client = new MqttClient(bootstrap, tls)) {
-                connection = new MqttClientConnection(client, events);
-
-                CompletableFuture<Boolean> connected = connection.connect(clientId, TEST_ENDPOINT, port, null, cleanSession, keepAliveMs, 0);
-                connected.get();
-                return true;
-            }
-        } catch (Exception ex) {
-            fail("Exception during connect: " + ex.toString());
-        }
-        return false;
-    }
-
-    void disconnect() {
-        disconnecting = true;
-        try {
-            CompletableFuture<Void> disconnected = connection.disconnect();
-            disconnected.get();
-        }
-        catch (Exception ex) {
-            fail("Exception during disconnect: " + ex.getMessage());
-        }
-
-    }
-
-    void close() {
-        connection.close();
-    }
-}
 
 public class MqttClientConnectionTest extends MqttClientConnectionFixture {
     public MqttClientConnectionTest() {
